Add files via upload
Testing for creating a function with parameters works fine.
Creating an int function throws an error. Processed as if setting an int type variable instead. Need to include in variable setting that it might be a function so after processing ID then it needs to check to see if there is a left paren or an equal sign. Paren means function, equals means variable.